aktos/requirements.txt
asgiref==3.8.1
coverage==7.4.4
Django==4.2.11
django-filter==24.2
djangorestframework==3.15.1
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.4.0
importlib_metadata==7.1.0
inflection==0.5.1
iniconfig==2.0.0
jedi==0.17.2
Markdown==3.6
packaging==24.0
parso==0.7.1
pluggy==1.4.0
psycopg2-binary==2.9.9
pyls-black==0.4.7
pytest==8.1.1
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.10.0
ujson==5.8.0
zipp==3.18.1
